refactor(lib): deduplicate resolveTasksBase and resolveWorktreeRoot into canonical helpers
Extract repeated ~30-line path resolution functions from per-step modules into
shared ticket-validation.js exports. All 25 *-next.js/*-phase-state.js files now
delegate to resolveTasksBaseWithFallback() (env > config > ~/worktrees/tasks)
and resolveWorktreeRoot() (git rev-parse --show-toplevel). Reduces duplication
without behavioral change; explicit fallback semantics clarify legacy compat.
