+++ scripts/workflows/work-cleanup/cleanup-next.js
@@ -28,27 +28,13 @@ try {
   logNextScriptEvent = () => {};
 }
 
+const { resolveTasksBaseWithFallback, resolveWorktreeRoot } = require('../lib/ticket-validation');
+
 let config;
 try {
   config = require('../lib/config');
 } catch {
   config = null;
-}
-
-function resolveTasksBase() {
-  return (
-    process.env.TASKS_BASE ||
-    (config && config.TASKS_BASE) ||
-    path.join(require('node:os').homedir(), 'worktrees', 'tasks')
-  );
-}
-
-function resolveWorktreeRoot() {
-  const r = spawnSync('git', ['rev-parse', '--show-toplevel'], {
-    encoding: 'utf8',
-    stdio: ['ignore', 'pipe', 'ignore'],
-  });
-  return r.status === 0 ? r.stdout.trim() : null;
 }
 
 function die(msg) {
@@ -174,7 +160,7 @@ function main(argv) {
 
   snapshotCompanionToken('cleanup-phase-state.js', ticket);
 
-  const tasksBase = resolveTasksBase();
+  const tasksBase = resolveTasksBaseWithFallback();
   const tasksDir = path.join(tasksBase, ticket);
   if (!fs.existsSync(tasksDir)) die(`tasks dir not found: ${tasksDir}`);
